refactor: share preview modal base for ads
